Use the embedding dropout
I don't know if the embedding dropout is defined by mistake (haven't check the tf implementation) but it's not used in the forward pass. This PR fixes it.
If it shouldn't be there at all, then the dropout and the `embd_pdrop` in the default config should probably be removed.